refactor runtime info
