+++ crates/cubecl-linalg/src/matmul/cmma/block_io/horizontal_block_check.rs
@@ -2,10 +2,9 @@ use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 
 use super::base::{BlockLoader, BlockWriter};
-use crate::matmul::cmma::{
-    base::{Dimensions, RuntimeCmmaInfo},
-    load_shared_memory::load_info::LoadInfo,
-};
+
+use crate::matmul::cmma::load_shared_memory::load_info::LoadInfo;
+use crate::matmul::cmma::runtime_info::{Dimensions, RuntimeCmmaInfo};
 
 pub(crate) struct HorizontalCheckBlockIO;
