+++ crates/cubecl-linalg/src/matmul/cmma/block_io/whole_block_check.rs
@@ -1,7 +1,5 @@
-use crate::matmul::cmma::{
-    base::{Dimensions, RuntimeCmmaInfo},
-    load_shared_memory::load_info::LoadInfo,
-};
+use crate::matmul::cmma::load_shared_memory::load_info::LoadInfo;
+use crate::matmul::cmma::runtime_info::{Dimensions, RuntimeCmmaInfo};
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
