+++ crates/cubecl-linalg/src/matmul/cmma/compute_loop/buffers_first.rs
@@ -1,13 +1,12 @@
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 
+use super::super::runtime_info::{Fragments, Ids, SharedMemories};
 use crate::matmul::cmma::{
-    base::{Fragments, SharedMemories},
     compute_loop::base::{
         get_smem_position_lhs, get_smem_position_rhs, load_tile_into_fragment, ComputeLoop,
     },
     config::ComptimeCmmaInfo,
-    ids::Ids,
 };
 
 pub(crate) struct AllBuffersFirstComputeLoop {}
