+++ crates/cubecl-linalg/src/matmul/cmma/load_shared_memory/tilewise.rs
@@ -1,8 +1,9 @@
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 
+use super::super::runtime_info::RuntimeCmmaInfo;
 use crate::matmul::cmma::{
-    base::RuntimeCmmaInfo, block_io::base::BlockLoader, config::ComptimeCmmaInfo,
+    block_io::base::BlockLoader, config::ComptimeCmmaInfo,
     load_shared_memory::base::get_tile_smem_index,
 };
