+++ crates/cubecl-linalg/src/matmul/cmma/write_output/reuse_smem.rs
@@ -1,8 +1,7 @@
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 
-use crate::matmul::cmma::base::RuntimeCmmaInfo;
-
+use super::super::runtime_info::RuntimeCmmaInfo;
 use super::{
     super::config::ComptimeCmmaInfo,
     base::{shared_memory_to_output, OutputWriter},
